pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[project]
name = "judge-extractor"
version = "0.1.0"
description = "Extract judge information from court decision PDFs"
requires-python = ">=3.11"
dependencies = [
    "pdfplumber>=0.10.0",
    "pdf2image>=1.16.0",
    "pymupdf>=1.24.0",
    "pytesseract>=0.3.10",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = ["pytest"]
